Now loading all assets into memory.
This should allow assets to be retreived by the frontend much more
quickly. I'm using to test how much of a it difference it can make.

Longer term there would be some other cache for asset metadata that
allows 1000 assets to be retreived on each call to S3, that would be
almost as good as building this in-memory cache.

The assets are definitely loaded more quickly into the frontend.

However the backend is very slow to load all assets.