Removed duplicates from search results.
The search algorithm was duplicating entries into the search results!

This was causing a bug with the layout algorithm. If an asset is
included more than once then it's layout will be the last position
that was allocated to it.

This results in gaps in the gallery where duplicates items are
automatically moved to the last position they hold in the gallery.