Separated "global index" and "display index".
The global index is the index within the reverse chronological
view. It can be used as an id for an asset within a collection.

The display index is the index of the asset within the current gallery.
This can be the same as the global index when assets are viewed in
reverse chronological order or it can be different when viewing the
results of a search.